\name{RCDT-package}
\alias{RCDT-package}
\alias{RCDT}
\docType{package}
\title{
\packageTitle{RCDT}
}
\description{
\packageDescription{RCDT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCDT}
\packageIndices{RCDT}
The \code{delaunay} function is the main function of this package. It can 
build a Delaunay triangulation of a set of 2D points, constrained or 
unconstrained. The constraints are defined by the \code{edges} argument.
}
\author{
\packageAuthor{RCDT}

Maintainer: \packageMaintainer{RCDT}
}
\keyword{ package }
