% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{delaunayArea}
\alias{delaunayArea}
\title{Area of Delaunay trianguation}
\usage{
delaunayArea(del)
}
\arguments{
\item{del}{an output of \code{\link{delaunay}}}
}
\value{
A number, the area of the Delaunay triangulation.
}
\description{
Computes the area of a Delaunay triangulation.
}
\examples{
library(RCDT)
set.seed(666)
library(uniformly)
square <- rbind(
  c(-1, 1), c(1, 1), c(1, -1), c(-1, -1)
)
pts <- rbind(square, runif_in_cube(8L, d = 2L))
del <- delaunay(pts)
delaunayArea(del)
}
