\name{RCDT-package}
\alias{RCDT-package}
\alias{RCDT}
\docType{package}
\title{
\packageTitle{RCDT}
}
\description{
\packageDescription{RCDT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCDT}
\packageIndices{RCDT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RCDT}

Maintainer: \packageMaintainer{RCDT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
